requirements.txt
py-staticmaps
pycairo
